feat(packages/sui-mono): Use new way to build commits message
